# projects/server/public/openapi.yaml
openapi: 3.1.0
info:
  title: Tool Vault API
  version: 1.0.0
  description: API for Tool Vault MCP gateway server (management, analystics, and proxy support)
servers:
  - url: http://localhost:3000/api/v1
    description: Tool Vault MCP Gateway
...
components:
  schemas:
    Meta:
      type: object
      properties:
        apiVersion:
          type: string
          enum: ["1.0"]
        status:
          type: integer
          description: HTTP status code
        message:
          type: string
          description: Status message
    Error:
      type: object
      properties:
        meta:
          $ref: '#/components/schemas/Meta'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /appSettings:
    get:
      tags:
        - Gateway
      summary: Get application settings
      responses:
        '200':
          description: Application settings retrieved successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Gateway
      summary: Update application settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Application settings updated successfully
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /analytics/dimensions:
    get:
      tags:
        - Analytics
      summary: Get dimension values for analytics
      description: Retrieves values for specified dimensions, supporting ID-based lookups, static values, and dynamic values from messages/alerts
      parameters:
        - name: dimension
          in: query
          required: true
          description: Dimensions to retrieve
          schema:
            type: array
            items:
              type: string
              enum:
                # Message dimensions
                - serverId
                - serverName
                - userId
                - clientId
                - clientType
                - payloadMethod
                - payloadToolName
                - sourceIP
                - sessionId
                # Alert dimensions
                - policyId
                - filterName
                - seen
                - severity
        - name: timeRange
          in: query
          required: false
          description: Time range for filtering (e.g. 24h, 7d)
          schema:
            type: string
        # Message filters
        - name: serverName
          in: query
          required: false
          description: Filter by server name
          schema:
            type: string
        - name: serverId
          in: query
          required: false
          description: Filter by server ID
          schema:
            type: integer
        - name: userId
          in: query
          required: false
          description: Filter by user ID
          schema:
            type: string
        - name: clientId
          in: query
          required: false
          description: Filter by client ID
          schema:
            type: integer
        - name: clientType
          in: query
          required: false
          description: Filter by client type
          schema:
            type: string
            enum:
              - cursor
              - roocode
              - windsurf
              - claudecode
              - vscode
              - generic
        - name: payloadMethod
          in: query
          required: false
          description: Filter by payload method
          schema:
            type: string
        - name: payloadToolName
          in: query
          required: false
          description: Filter by payload tool name
          schema:
            type: string
        # Alert filters
        - name: policyId
          in: query
          required: false
          description: Filter by policy ID
          schema:
            type: integer
        - name: filterName
          in: query
          required: false
          description: Filter by filter name
          schema:
            type: string
        - name: seen
          in: query
          required: false
          description: Filter by seen status
          schema:
            type: boolean
        - name: severity
          in: query
          required: false
          description: Filter by severity level
          schema:
            type: integer
            enum: [1, 2, 3, 4, 5]
      responses:
        '200':
          description: Successfully retrieved dimension values
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    additionalProperties:
                      type: array
                      items:
                        type: object
                        properties:
                          value:
                            type: string
                            description: The dimension value
                          label:
                            type: string
                            description: The human-readable label for the value
                        required:
                          - value
                          - label
                  query:
                    type: object
                    properties:
                      dimensions:
                        type: array
                        items:
                          type: string
                      timeRange:
                        type: object
                        properties:
                          start:
                            type: string
                            format: date-time
                          end:
                            type: string
                            format: date-time
                      filters:
                        type: object
                        properties:
                          serverName:
                            type: string
                          serverId:
                            type: integer
                          userId:
                            type: string
                          clientId:
                            type: integer
                          clientType:
                            type: string
                          payloadMethod:
                            type: string
                          payloadToolName:
                            type: string
                          sourceIP:
                            type: string
                          sessionId:
                            type: string
                          policyId:
                            type: integer
                          filterName:
                            type: string
                          seen:
                            type: boolean
                          severity:
                            type: integer
                required:
                  - data
                  - query
        '400':
          description: Bad request - missing required parameters or invalid values
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # Client-Server Relationship Endpoints
...